moved leaves
